Add cleanup logic to runAgent
